# demos/hypermedia-server/static/meta/openapi.json
{
  "info": {
    "title": "Hypermedia Demo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/todos": {
      "get": {
        "operationId": "getTodos",
        "parameters": [
          {
            "description": "Opaque pagination cursor returned by a previous response",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "urn:schema:hal:nb.TodoCollection:1.0.0"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "urn:schema:nb.TodoCollection:1.0.0"
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": ["Todo"]
      },
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "urn:schema:nb.CreateTodo:1.0.0"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "urn:schema:svc.CommandSuccessResponse:1.0.0"
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": ["Todo"]
      }
    },
    "/api/todos/{id}": {
      "get": {
        "operationId": "getTodoById",
        "parameters": [
          {
            "description": "Todo identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "urn:schema:hal:nb.Todo:1.0.0"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "urn:schema:nb.Todo:1.0.0"
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": ["Todo"]
      }
    },
...
  },
...
}